+++ hpg-bigdata-tools/src/main/java/org/opencb/hpg/bigdata/tools/alignment/stats/ReadAlignmentSortMR.java
@@ -49,9 +49,9 @@ public class ReadAlignmentSortMR {
             ChunkKey newKey;
             LinearAlignment la = (LinearAlignment) key.datum().getAlignment();
             if (la == null) {
-                newKey = new ChunkKey(new String("*"), (long) 0);
+                newKey = new ChunkKey(new String("*"), 0);
             } else {
-                newKey = new ChunkKey(la.getPosition().getReferenceName().toString(), la.getPosition().getPosition());
+                newKey = new ChunkKey(la.getPosition().getReferenceName().toString(), la.getPosition().getPosition().intValue());
             }
             context.write(newKey, new AvroValue<>(key.datum()));
         }
